LLONG_MAX to SIZE_MAX

+++ src/cjson.c
@@ -37,6 +37,7 @@
 #pragma warning (disable : 4001)
 #endif
 
+#include "iperf_config.h"
 #include <string.h>
 #include <stdio.h>
 #include <math.h>
@@ -61,7 +62,6 @@
 #endif
 
 #include "cjson.h"
-#include "iperf_config.h"
 
 /* define our own boolean type */
 #ifdef true
@@ -464,9 +464,9 @@ static unsigned char* ensure(printbuffer * const p, size_t needed)
         return NULL;
     }
 
-    if (needed > LLONG_MAX)
+    if (needed > SIZE_MAX)
     {
-        /* sizes bigger than LLONG_MAX are currently not supported */
+        /* sizes bigger than SIZE_MAX are currently not supported */
         return NULL;
     }
 
@@ -481,12 +481,12 @@ static unsigned char* ensure(printbuffer * const p, size_t needed)
     }
 
     /* calculate new buffer size */
-    if (needed > (LLONG_MAX / 2))
+    if (needed > (SIZE_MAX / 2))
     {
-        /* overflow of int, use LLONG_MAX if possible */
-        if (needed <= LLONG_MAX)
+        /* overflow of int, use SIZE_MAX if possible */
+        if (needed <= SIZE_MAX)
         {
-            newsize = LLONG_MAX;
+            newsize = SIZE_MAX;
         }
         else
         {
